feat: glfw edit_text_dialog
