+++ library/lib/platforms/glfw/glfw_ime.cpp
@@ -21,6 +21,7 @@ limitations under the License.
 #include <borealis/platforms/glfw/glfw_ime.hpp>
 #include <borealis/views/dialog.hpp>
 #include <borealis/views/label.hpp>
+#include <borealis/views/edit_text_dialog.hpp>
 #include <codecvt>
 #include <cstring>
 #include <iostream>
@@ -28,190 +29,6 @@ limitations under the License.
 
 namespace brls
 {
-
-const std::string editTextDialogXML = R"xml(
-    <brls:Box
-        id="brls/container"
-        width="auto"
-        height="auto"
-        axis="column"
-        justifyContent="flexStart"
-        alignItems="center"
-        focusable="true"
-        hideHighlight="true"
-        backgroundColor="@theme/brls/backdrop">
-
-        <brls:Label
-            id="brls/dialog/header"
-            fontSize="24"
-            marginTop="50"
-            marginBottom="30"
-            textColor="#FFFFFF"/>
-
-        <brls:Box
-            id="brls/dialog/applet"
-            width="720"
-            cornerRadius="4"
-            alignItems="flexEnd"
-            axis="column"
-            backgroundColor="@theme/brls/background">
-
-            <brls:Label
-                id="brls/dialog/label"
-                grow="1"
-                width="680"
-                margin="20"
-                autoAnimate="false"
-                verticalAlign="top"/>
-
-            <brls:Label
-                id="brls/dialog/count"
-                width="680"
-                horizontalAlign="right"
-                fontSize="18"
-                textColor="@theme/brls/text_disabled"
-                marginRight="30"
-                marginBottom="10"/>
-
-            <brls:Hints
-                allowAButtonTouch="true"
-                addBaseAction="false"
-                marginBottom="20"
-                marginRight="10"
-                width="auto"
-                height="auto"/>
-
-        </brls:Box>
-
-    </brls:Box>
-)xml";
-
-class EditTextDialog : public Box
-{
-  public:
-    EditTextDialog()
-    {
-        this->inflateFromXMLString(editTextDialogXML);
-
-        // submit text
-        this->registerAction(
-            "hints/submit"_i18n, BUTTON_START, [this](...)
-            {
-                Application::popActivity(TransitionAnimation::FADE, [this](){
-                        this->summitEvent.fire();
-                    });
-                return true; });
-        this->registerAction(
-            "hints/ok"_i18n, BUTTON_A, [this](...)
-            {
-                Application::popActivity(TransitionAnimation::FADE, [this](){
-                        this->summitEvent.fire();
-                });
-                return true; },
-            true);
-
-        // cancel input
-        this->registerAction(
-            "hints/back"_i18n, BUTTON_BACK, [this](...)
-            {
-                Application::popActivity(TransitionAnimation::FADE, [this](){
-                        this->cancelEvent.fire();
-                    });
-
-                return true; });
-
-        this->init = true;
-    }
-
-    void open()
-    {
-        Application::pushActivity(new Activity(this));
-    }
-
-    void setText(const std::string& value)
-    {
-        this->content = value;
-        this->updateUI();
-    }
-
-    void setHeaderText(const std::string& value)
-    {
-        this->header->setText(value);
-    }
-
-    void setHintText(const std::string& value)
-    {
-        if (value.empty())
-        {
-            this->hint = "hints/input"_i18n;
-        }
-        else
-        {
-            this->hint = value;
-        }
-        this->updateUI();
-    }
-
-    void setCountText(const std::string& value)
-    {
-        this->count->setText(value);
-    }
-
-    bool isTranslucent() override
-    {
-        return true;
-    }
-
-    void onLayout() override
-    {
-        if (!init)
-            return;
-        this->layoutEvent.fire(Point { this->label->getX(),
-            this->label->getY() + this->label->getHeight() });
-    }
-
-    Event<Point>* getLayoutEvent()
-    {
-        return &this->layoutEvent;
-    }
-
-    Event<>* getCancelEvent()
-    {
-        return &this->cancelEvent;
-    }
-
-    Event<>* getSubmitEvent()
-    {
-        return &this->summitEvent;
-    }
-
-    void updateUI()
-    {
-        if (content.empty())
-        {
-            label->setTextColor(Application::getTheme().getColor("brls/text_disabled"));
-            label->setText(hint);
-        }
-        else
-        {
-            label->setTextColor(Application::getTheme().getColor("brls/text"));
-            label->setText(content);
-        }
-    }
-
-  private:
-    std::string content;
-    std::string hint;
-    Event<Point> layoutEvent;
-    Event<> cancelEvent, summitEvent;
-    bool init = false;
-
-    BRLS_BIND(brls::Label, header, "brls/dialog/header");
-    BRLS_BIND(brls::Label, label, "brls/dialog/label");
-    BRLS_BIND(brls::Label, count, "brls/dialog/count");
-    BRLS_BIND(brls::Box, container, "brls/container");
-};
-
 static int currentIMEStatus = GLFW_FALSE;
 #define MAX_PREEDIT_LEN 128
 static char preeditBuf[MAX_PREEDIT_LEN] = "";
